fix(poke): align superuser admin with coding rules
