+++ front-api/middlewares/utils.ts
@@ -153,7 +153,7 @@ export const unhandledErrorHandler: ErrorHandler = (err, ctx) => {
     {
       error: {
         type: "internal_server_error",
-        message: "An unexpected internal error occurred.",
+        message: `Unhandled internal server error: ${error.message}`,
       },
     },
     500
