+++ front-api/routes/poke/superusers/[userId]/index.ts
@@ -1,0 +1,9 @@
+import { pokeApp } from "@front-api/middlewares/ctx";
+
+import superuser from "./superuser";
+
+const app = pokeApp();
+
+app.route("/superuser", superuser);
+
+export default app;
